Refactor dependencies so that CLI dependencies are optional
Add the "cli" feature that enables dependencies that are only used by
the CLI, this makes the library compilation smaller and easier to build
for some targets (such as android) as it eliminates TLS and such.

After this change to build the CLI you must run:
cargo build --bin esptool --features cli

cargo build with no arguments will only build the library.