Remove serde-hex crate due to save / load conflict
When using serde-hex, the vid and pid are saved as:

    [connection]

    [[usb_device]]
    vid = "303a"
    pid = "1001"

in the ~/.config/espflash/espflash.toml file.

However, upon loading, we see an error message:

Error:
  × TOML parse error at line 4, column 7
  │   |
  │ 4 | vid = "303a"
  │   |       ^^^^^^
  │ invalid type: string "303a", expected a borrowed byte array
  │

The serde-hex source code hasn't been updated in 4 years, and
is not apparently compatible with the toml crate.

Storing the vid and pid as decimal integers is not ideal, but
it at least works correctly.