Always resolve package_id from metadata when finding bootloader and partition table
The package ID format was officially documented as an opaque identifier
and the only thing we were allowed to do with them is look them up in
the cargo metadata.  Rust 1.77 switched the package ID field to use the
publicly documented "Package ID Specification" format instead, which
means that the old logic no longer works.

Signed-off-by: Johannes Löthberg <johannes.loethberg@elokon.com>