Write the image to the correct target app partition
This updates the code to honor the --target-app-partition flag,
and write the application image to the specified partition.  Previously
the code always wrote the image to the "factory" app partition if one
was present, regardless of the partition specified with the
`--target-app-partition` flag.  (It seems like this behavior was likely
broken in d886d33a69, which updated the code to always write to the
factory partition again.)

I confirmed that using `espflash flash --target-app-partition ota_0` now
causes espflash to write the image to the `ota_0` partition.  Note that
this leaves the `factory` partition and any `ota_data` partition
unchanged, so the bootloader will still boot from the factory partition
by default, but this makes it possible to write new app images to the
OTA partitions for testing purposes.